refactor(tui): split form model responsibilities
